+++ service/client_info.go
@@ -23,6 +23,7 @@ import (
 	api "github.com/polarismesh/polaris-server/common/api/v1"
 	"github.com/polarismesh/polaris-server/common/model"
 	"github.com/polarismesh/polaris-server/common/utils"
+	"go.uber.org/zap"
 )
 
 var (
@@ -34,6 +35,57 @@ var (
 		"version": {},
 	}
 )
+
+func (s *Server) checkAndStoreClient(ctx context.Context, req *api.Client) *api.Response {
+	clientId := req.GetId().GetValue()
+	var needStore bool
+	client := s.caches.Client().GetClient(clientId)
+	var resp *api.Response
+	if nil == client {
+		needStore = true
+	} else {
+		needStore = !clientEquals(client.Proto(), req)
+	}
+	if needStore {
+		client, resp = s.createClient(ctx, req)
+	}
+
+	if resp != nil {
+		if resp.GetCode().GetValue() != api.ExistedResource {
+			return resp
+		}
+	}
+
+	return s.HealthServer().ReportByClient(context.Background(), req)
+}
+
+func (s *Server) createClient(ctx context.Context, req *api.Client) (*model.Client, *api.Response) {
+
+	if namingServer.bc == nil || !namingServer.bc.CreateInstanceOpen() {
+		return nil, nil
+	}
+	return s.asyncCreateClient(ctx, req) // 批量异步
+}
+
+// 异步新建客户端
+// 底层函数会合并create请求，增加并发创建的吞吐
+// req 原始请求
+// ins 包含了req数据与instanceID，serviceToken
+func (s *Server) asyncCreateClient(ctx context.Context, req *api.Client) (*model.Client, *api.Response) {
+	rid := utils.ParseRequestID(ctx)
+	pid := utils.ParsePlatformID(ctx)
+	future := s.bc.AsyncRegisterClient(req)
+	if err := future.Wait(); err != nil {
+		log.Error("[Server][ReportClient] async create client", zap.Error(err), ZapRequestID(rid),
+			ZapPlatformID(pid))
+		if future.Code() == api.ExistedResource {
+			req.Id = utils.NewStringValue(req.GetId().GetValue())
+		}
+		return nil, api.NewClientResponse(future.Code(), req)
+	}
+
+	return future.Client(), nil
+}
 
 // CreateInstances create one instance
 func (s *Server) GetReportClients(ctx context.Context, query map[string]string) *api.BatchQueryResponse {
@@ -93,4 +145,46 @@ func client2Api(client *model.Client) *api.Client {
 	}
 	out := client.Proto()
 	return out
+}
+
+func clientEquals(client1 *api.Client, client2 *api.Client) bool {
+	if client1.GetId().GetValue() != client2.GetId().GetValue() {
+		return false
+	}
+	if client1.GetHost().GetValue() != client2.GetHost().GetValue() {
+		return false
+	}
+	if client1.GetVersion().GetValue() != client2.GetVersion().GetValue() {
+		return false
+	}
+	if client1.GetType() != client2.GetType() {
+		return false
+	}
+	if client1.GetLocation().GetRegion().GetValue() != client2.GetLocation().GetRegion().GetValue() {
+		return false
+	}
+	if client1.GetLocation().GetZone().GetValue() != client2.GetLocation().GetZone().GetValue() {
+		return false
+	}
+	if client1.GetLocation().GetCampus().GetValue() != client2.GetLocation().GetCampus().GetValue() {
+		return false
+	}
+	if len(client1.Stat) != len(client2.Stat) {
+		return false
+	}
+	for i := 0; i < len(client1.Stat); i++ {
+		if client1.Stat[i].GetTarget().GetValue() != client2.Stat[i].GetTarget().GetValue() {
+			return false
+		}
+		if client1.Stat[i].GetPort().GetValue() != client2.Stat[i].GetPort().GetValue() {
+			return false
+		}
+		if client1.Stat[i].GetPath().GetValue() != client2.Stat[i].GetPath().GetValue() {
+			return false
+		}
+		if client1.Stat[i].GetProtocol().GetValue() != client2.Stat[i].GetProtocol().GetValue() {
+			return false
+		}
+	}
+	return true
 }
